Release 0.1.0
* Bump gradle version

* XCFramework support

* Bump core version

* 0.1.0
