+++ .circleci/config.yml
@@ -12,7 +12,7 @@ jobs:
 
   build-ios:
     macos:
-      xcode: "10.0"
+      xcode: "12.0"
 
     steps:
       - checkout
@@ -43,7 +43,7 @@ jobs:
 
   unit-test:
     macos:
-      xcode: "10.0"
+      xcode: "12.0"
 
     steps:
       - checkout
@@ -77,7 +77,7 @@ jobs:
 
   build-sample-app-ios:
     macos:
-      xcode: "10.0"
+      xcode: "12.0"
 
     steps:
       - checkout
