+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@adobe/react-native-acpplaces",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "Adobe Experience Platform support for React Native apps.",
   "homepage": "https://aep-sdks.gitbook.io/docs/",
   "license": "Apache-2.0",
