+++ sample/ACPPlacesSample/package.json
@@ -7,8 +7,9 @@
     "test": "jest"
   },
   "dependencies": {
-    "@adobe/react-native-acpcore": "^1.1.2",
+    "@adobe/react-native-acpcore": "github:adobe/react-native-acpcore#xcframeworks-release",
     "@adobe/react-native-acpplaces": "^0.0.1",
+    "graceful-fs": "4.2.2",
     "@babel/runtime": "^7.0.0",
     "babel-cli": "^6.26.0",
     "babel-eslint": "^8.2.2",
